+++ coursedata.js
@@ -1,0 +1,10 @@
+// version: 0.0.5_250104-1051
+coursename='Demonstration Course for CourseMakerHTML <br/>by Ian Low';
+coursedesc='This is a test course demonstrating how a course could look like.';
+coursefldr='democourse';
+sectionfldr='1starter';
+sectionlist="\
+1. Starter|1starter|Introductory section for all starters%%%\
+2. Basic|2basic|Basic lessons teaching the fundamentals%%%\
+3. Intermediate|3intermediate|Intermediate lessons teaching the more complex concepts%%%\
+4. Advanced|4advanced|Advanced lessons teaching the more advanced concepts";