+++ democourse/2basic/part02.js
@@ -1,19 +1,18 @@
-part02="{Demonstration Course for CourseMakerHTML by Fred Bloggs}\
-\n\
-\n\
-[Part Two - Introduction]\n\
-\n\
-<b>Background</b><br/><br/>\
-##1 This is an introduction on how to use this application<br/>\
-##2 Here is line 2<br/>\
-##3 Here is line 3<br/>\
-\
-\n\
-\n\
+part02="\
+##1 Fill in the blanks:|The week begins on [Monday] followed by [Tuesday]. \n\
 \
 \
-## <b>Further information</b><br/>\
-##1 This is further information on how to use this application\n\
-";
-
-alert(part01);
+\
+##2 The week starts on...|Thursday|[Monday]|Saturday\n\
+\
+\
+\
+##3 Which of the following are weekdays?|[Thursday]|[Monday]|Saturday \n\
+\
+\
+\
+##4 Which of the following are weekdays? %%% Unfortunately your answer was incorrect.|[Thursday] |[Monday] | Saturday \n\
+\
+\
+\
+Which of the following are weekdays?%%% Unfortunately your answer was incorrect. You will now be directed to page 8 of part 2 of the Starter section. %%% 1starter %%2 %%8 |[Thursday] |[Monday] | Saturday \n";