+++ democourse/3intermediate/part01.js
@@ -1,0 +1,1 @@
+part01="test3|[a+b]|b|c\n";