+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 
 ## NOTES
 
+## 0.0.5_250104-1051 - Added usage information in the demo course and use of coursedata.js for the definition of a specific course.
+-  Added usage information in the demo course and use of coursedata.js for the definition of a specific course.
+
 ## 0.0.4_250103-0242 - Additional enhancements
 -  Additional enhancements